Move readable jobs to hook
